+++ handlers/handlersMain_test.go
@@ -1,13 +1,20 @@
 package handlers_test
 
 import (
+	"flag"
 	"os"
 	"testing"
 
 	"github.com/accedian/adh-gather/logger"
 )
 
+var (
+	metricsIntegrationTests = flag.Bool("metrics", false, "Run metrics api integration tests")
+)
+
 func TestMain(m *testing.M) {
+
+	flag.Parse()
 
 	err := setupTestDatastore()
 	if err != nil {